refactor: App component 분리
